(tes) Improve spanning graph creation.
The step would probably be fused with the next stage (add edges) in the future, because of a low diversity at low connectivity (where it's unavoidable) that I want to avoid at high connectivity (where it is avoidable, but needs that refactoring).

Also some adjustments to the tablegen. I have some new ideas about tablegens, but they need even bigger refactors.

Sometimes you have to live with bad stuff for a while.