(tes, wip) Hexagon + Juxposition
The combination proved to be much harder than sqaure grid. Or in other words,
the square grid does not differentiate between a few conceptually different things
such as "part A looks the same as part B", "part A and B are mirrors" and "part A
and B have a reflectional axis in between".

Working on the hexagonal grid forces me to think clearly about each situation.